refactor: 테스트 코드 클래스 분리
