+++ src/main/java/nextstep/ladder/domain/Ladder.java
