+++ src/main/java/nextstep/ladder/domain/LineGenerateStrategy.java
