+++ src/main/java/nextstep/ladder/domain/Point.java
