+++ src/main/java/nextstep/ladder/domain/RandomLineGenerateStrategy.java
